+++ src/ladb_opencutlist/ruby/controller/settings_controller.rb
@@ -22,11 +22,13 @@ module Ladb::OpenCutList
     def dialog_settings_command(settings)
 
       # Check settings
+      language = settings['language']
       width = settings['width']
       height = settings['height']
       top = settings['top']
       left = settings['left']
 
+      Plugin.instance.set_language(language, true)
       Plugin.instance.dialog_set_size(width, height, true)
       Plugin.instance.dialog_set_position(left, top, true)